// frontend/sdk/models/StreakQuestionDto.ts
export interface StreakQuestionDto {
    /**
     * 
     * @type {string}
     * @memberof StreakQuestionDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof StreakQuestionDto
     */
    question: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof StreakQuestionDto
     */
    options: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof StreakQuestionDto
     */
    correctAnswerIndex: number;
    /**
     * 
     * @type {string}
     * @memberof StreakQuestionDto
     */
    category: string;
    /**
     * 
     * @type {string}
     * @memberof StreakQuestionDto
     */
    context?: string;
    /**
     * 
     * @type {boolean}
     * @memberof StreakQuestionDto
     */
    isActive: boolean;
    /**
     * 
     * @type {string}
     * @memberof StreakQuestionDto
     */
    addedBy: string;
    /**
     * 
     * @type {Date}
     * @memberof StreakQuestionDto
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof StreakQuestionDto
     */
    updatedAt: Date;
}